move env parsingt to lib shared
