+++ packages/loudness/src/config.ts
@@ -4,7 +4,7 @@ dotenv.config();
 // load loudness.config.ts or loudness.config.js if present - otherwise use defaults or env - .ts needs to be compiled to .js
 //if(fs.existsSync('../loudness.config.ts')) {
 //import configfile from '../loudness.config.ts';
-import { type ShareInfo } from '@airlookjs/shared';
+import { parseIntEnv, type ShareInfo } from '@airlookjs/shared';
 
 export interface LoudnessConfig {
   environment: string;
@@ -14,33 +14,7 @@ export interface LoudnessConfig {
   port: number;
 };
 
-// parse bools from env safely
-/*const parseBoolEnv = (env: string | undefined, defaultValue: boolean) => {
-	if (env === undefined) {
-		return defaultValue
-	}
-	if (env === 'true') {
-		return true
-	}
-	if (env === 'false') {
-		return false
-	}
-	return defaultValue
-}*/
-
-// parse ints from env safely
-const parseIntEnv = (env: string | undefined, defaultValue: number) => {
-	if (env === undefined) {
-		return defaultValue
-	}
-	const parsed = parseInt(env)
-	if (isNaN(parsed)) {
-		return defaultValue
-	}
-	return parsed
-}
-
-export const config: LoudnessConfig = {
+export const config: LoudnessConfig = {	
 	environment: process.env.NODE_ENV ?? 'development',
 	version: process.env.npm_package_version ?? 'dev',
 	port: parseIntEnv(process.env.PORT, 3000),
@@ -48,4 +22,4 @@ export const config: LoudnessConfig = {
 	shares: []
 };
 
-export const LOUDNESS_CMD = process.env.LOUDNESS_CMD ?? './bin/loudness';
+export const LOUDNESS_CMD = './bin/loudness';
