+++ packages/mediainfo/src/config.ts
@@ -1,39 +1,4 @@
-import { type ShareInfo } from '@airlookjs/shared';
-
-// parse bools from env safely
-/*const parseBoolEnv = (env, defaultValue) => {
-	if (env === undefined) {
-		return defaultValue
-	}
-	if (env === 'true') {
-		return true
-	}
-	if (env === 'false') {
-		return false
-	}
-	return defaultValue
-}*/
-
-// parse ints from env safely
-const parseIntEnv = (env: string | undefined, defaultValue: number) => {
-	if (env === undefined) {
-		return defaultValue
-	}
-	const parsed = parseInt(env)
-	if (isNaN(parsed)) {
-		return defaultValue
-	}
-	return parsed
-}
-
-/*const sharedConfig = getSharedConfig({
-	shares: {
-		agis: {
-			mount: process.env.SHARE_AIRLOOK_MOUNT || '/mnt/agis-store',
-			cached: parseBoolEnv(process.env.SHARE_AIRLOOK_CACHED, true)
-		}
-	},
-})*/
+import { parseIntEnv, type ShareInfo } from '@airlookjs/shared';
 
 // TODO: parse config file to docker image to set up shares
