+++ packages/scenedetect/src/config.ts
@@ -1,4 +1,4 @@
-import { ShareInfo } from '@airlookjs/shared';
+import { parseIntEnv, type ShareInfo } from '@airlookjs/shared';
 import dotenv from 'dotenv';
 dotenv.config();
 
@@ -10,32 +10,6 @@ export interface SceneDetectConfig {
 	version: string;
 	port: number;
 }
-
-// parse bools from env safely
-/*const parseBoolEnv = (env: string | undefined, defaultValue: boolean) => {
-	if (env === undefined) {
-		return defaultValue;
-	}
-	if (env === 'true') {
-		return true;
-	}
-	if (env === 'false') {
-		return false;
-	}
-	return defaultValue;
-};*/
-
-// parse ints from env safely
-const parseIntEnv = (env: string | undefined, defaultValue: number) => {
-	if (env === undefined) {
-		return defaultValue;
-	}
-	const parsed = parseInt(env);
-	if (isNaN(parsed)) {
-		return defaultValue;
-	}
-	return parsed;
-};
 
 export const config: SceneDetectConfig = {
 	environment: process.env.NODE_ENV ?? 'development',
